common/requirements/base.txt
Django==1.10.6
django-cors-headers==2.0.0
djangorestframework==3.6.2
django-redis==4.7.0
django-extensions==1.7.5
django-filter==1.0.1
django-mysql==1.1.0
feedparser==5.2.1
mysqlclient==1.3.9
Pillow==4.0.0
pytz==2016.10
pydocs==0.2
celery==4.0.2
django-storages==1.5.2
boto3==1.4.3
twilio== 5.6.0
phonenumbers==8.1.0
redis==2.10.5
stream-python==2.3.9
av==0.3.3
